deepseek-chat-demo/requirements.txt
flask
python-dotenv
requests
pypdf
python-docx
openpyxl
